+++ docs/changelog.rst
@@ -10,7 +10,8 @@ Raster Vision 0.10.0
 Features
 ^^^^^^^^^^^^
 
-- Add confusion matrix to eval for semantic segmentation `#786 <https://github.com/azavea/raster-vision/pull/786>`__
+- Add confusion matrix as metric for semantic segmentation `#788 <https://github.com/azavea/raster-vision/pull/788>`__
+- Add predict_chip_size as option for semantic segmentation `#786 <https://github.com/azavea/raster-vision/pull/786>`__
 - Handle "ignore" class for semantic segmentation `#783 <https://github.com/azavea/raster-vision/pull/783>`__
 
 Bug Fixes